backend/requirements.txt
fastapi
uvicorn
deepface
pillow
python-multipart
PyJWT
python-jose[cryptography]
httpx
pytest
pytest-asyncio
tf-keras
cryptography
redis==5.0.1 